[Login] Fix static calls to User class to be non-static (#5247)
The password expiry code was making static calls to updatePassword and passwordChanged in the User class which are non-static methods. This created errors such as PHP Fatal error: Uncaught Error: Using $this when not in object context, making it impossible to update an expired password due to an oversight in #4411.