[Script] Add script to generate fake candidates (#5300)
This adds a script used to generate a large number of candidates
for #5299. The candidates are not realistic or even necessarily
valid, but only intended to see how the candidate_list data table
holds up with a large number of candidates (I tested with up to
20,000.)

This needs a lot of work to be made robust enough to be useable
for general purpose testing and is unlikely to be maintained, so
is added directly to the "deprecated" directory.