[imaging uploader] Automatically populate PSCID, CandID & visit label based on the filename upon upload (#5306)
Modified the behaviour of the upload tab of the imaging uploader module. PSCID, CandID and Visit Label fields are now automatically populated based on the name of the selected file to be uploaded.

The DICOM header and file name still must match as a form of double data entry/verification.