Fix edge-case that resulted in 500 error (#5320)
There was a slight issue with the logic during the install process that swallowed a DB exeption in favour of returning false; however, the return value of this function was never checked so the code kept trying to function without a DB connection. This could give a 500 error in some cases, making the install look messy and hard to debug.