[Configuration] display radio buttons when boolean option does not have a default option (#5592)
This fixes the rendering of boolean fields in the configuration module. without this change a boolean field without a default option set (field defined in ConfigSettings but no entry in Config table) renders as a text entry in the configuration module. it should render as true/false radio buttons with non selected with this change.