[configurations] Allow for both 0/1 or true/false to be used (#5605)
This restores a behaviour from #5592 while keeping the bugfix.

Previous behaviour was intentionally maintaining 1/0 values when previously saved value was 1/0 and true/false values when previously saved value was true/false. That is due to a lack of standard when coding with boolean values within modules.

The PR linked above converted all values to 1/0 automatically which is problematic because some code sections will be expecting a true or false to be returned from the getSetting('something') call.