[JSX] Enabled Data Mapping for CSV Downloads (#5291)
Because the getFormatedCell function often both formats AND maps data, there was a developing issue where the downloaded data in the form of CSV did not represent what was shown in the datatable, but rather the initial data before mapping - therefore downloading keys rather than values.

This fixes this by allowing developers to disentangle mapping and formatting. A function can be created which checks the value of the parameter column and then changes the value correspondingly. This function can then be passed as a prop to the FilterableDataTable component which will take care of the rest.