[brainbrowser] fix multiple volumes display bug (#5677)
When trying to load more than one volume in BB via the imaging browser, only the first volume got displayed. When we refreshed the page, then the second volume would be displayed and so on.

This fixes that bug by using a promise chain to not display until all images are loaded.

Fixes #5474.