[ReadTheDocs] Add existing wiki files to the navigation tree  (#5844)
Adds some headings and basic files to our read the docs page. These reflect the directory structure on the back-end. Note that this only includes a few templates and does not contain the content of the archived wiki pages.

This organization isn't necessarily the final layout for the documentation. It is meant to provide a helpful visual and help us to get moving on our big documentation revamp.

This file should also give an indication to other members of the team how to add new pages to ReadTheDocs when making their own documentation PRs.