[Meta] Add CODE_OF_CONDUCT.md
This adds a CODE_OF_CONDUCT.md file with the goal of being as
non-politicized as possible. It defers to MCIN and McGill University
codes and policies (which presumably have been vetted by HR and legal
departments applicable to us) with a vague paragraph imploring respect
and good faith for people who are not part of those organizations.