[Dashboard] Fix study progression performance
For long running studies the "study progression" widget on the
dashboard was doing an excessive number of queries on the database
(one per site per month of the study having been running) which
was causing the dashboard to take an excessive amount of time to
load.

This replaces it with a single query using a GROUP BY, and then
massages the data in PHP to the same format.