[Dashboard/Statistics] Move dashboard ajax endpoints into the statistics module (#5900)
This moves the ajax calls from the dashboard module into the more appropriate
statistics module for the graph endpoints. It also converts some inexplicable
"post" requests into "get" requests but otherwise behaviour should remain
unchanged.

This was extracted from #5896 and is sent as a separate PR in order to simplify
that PR and make it easier to rebase data/permission related PRs which would
otherwise conflict.