[API/Router] Move UnsupportedMediaType to the json folder
UnsupportedMediaType is a type of JSON response with a specific
HTTP response code. The API was instantiating it as if it was
in the LORIS\Http\Response\JSON namespace (as all the other
JSON specific response classes are) but the file was in
LORIS\Http\Response. This moves the file to the correct
namespace.