Document LORIS automated testing (#5859)
Adds a document that gives a high-level overview of our TravisCI pipeline.

This is intended as a general explainer of each of the tools we use. It will not include a list of each and every rule enforced by different tools nor provide example issues and solutions. Instead, links will be provided to the repositories of those tools as they'll be able to provide better detail than we can.

Resolves #5819