Remove unnecessary (and potentially incorrect comment) and PHPMD rule (#5936)
This removes a comment that was mistakenly merged in #4404. It also
removes a design rule from the config that should never have been
added, but only got included because PHPMD's config file is based
on excluding rules that you don't want, and not including rules
that you want, and the rule was snuck in in a PHPMD upgrade.