Remove BaseURL config setting (#6019)
The \LORIS\BaseRouter was already calculating the BaseURL of the incoming request and adding it to the request object. Unfortunately, most places in LORIS are not using it. Most instances would need to be refactored to get access to it since they're often deeper in the code and using \NDB_Factory->settings()->getBaseURL() at places with no direct access to the PSR7 request instance. This updates the code to have the router pass the calculated URL to the factory singleton instance which injects it into the Settings object.

There were a very straggling instances to the (even older) $config->getSetting('www')['url'] style of getting the baseURL which had to be updated to use the factory.

Fixes #5395, #6007.