Fix "Undefined index: filter" errors in NDB_Menu_Filter
Fix errors of the type "Undefined index: filter" when accessing
some NDB_Menu_Filter pages. This prevents some modules from loading
because the notice results in the JSON returned by LORIS being
invalid.

(This PR is a duplicate of #6011 but with the requested changes
added, since some modules currently can't be loaded without the
fix.)