[Core] Do not crash when there are Active invalid modules (#6063)
When there is an active module installed in the module table
that does not exist LORIS crashes with an exception while trying
to build the menus. This catches the exception and logs it to the
error log instead of dying.

While the 'Active' invalid module indicates a configuration error
that should be fixed, it can often happen in development environments
when switching branches, so instead we just log it in the error log.