[CandID] Fix validation of CandID object types (#6084)
The CandID object type was verifying that a CandID had 6 integers
in it. It did not validate that it only had 6 integers, and as a
result was considering everything that started with 6 integers valid
even if there was trailing invalid characters. This fixes the regular
expression so that CandID are 6 digit integers, and only 6 digit integers.