[candidate_parameters] Add DoD Feature (#4929)
This adds a new tab to candidate parameters which can be used to record a candidate's date of death. The PR also adds support for this new feature in instrument class. If the date of administration of an instrument is after the date of death, the candidate's age at death will be displayed rather than age at time of administration.