[candidate_profile] Add behavioural data widget
This adds a "Behavioural Data" widget to the candidate_profile
page for anyone who has access to the instruments module. The
Behavioural Data consists of a list of (initially collapsed)
visits which expand to a list of instruments and the data entry
percentage which is complete. Links are provided to the instrument
for data entry, the visit, and the conflict resolver (if applicable).