[JSX] Fix ESLint errors in <Loader /> component (#6109)
The component is used multiple places, so generates numerous
warnings when running `make dev`. It's also very simple, so
easy to fix.

This also converts the component from a class to a function,
because that seems to be more in-line with the direction that
React is taking with React hooks, and there is no external
state.