[Core] Add config option for projects to specify how they want dates to be displayed (#5004)
Adds a config option to allow projects to specify how they want dates to be displayed in LORIS.

This only adds a quick examples of using this Config in the issue_tracker. It is by no means comprehensive.

This will also only take place in "PHP contexts" and not within JS date elements as these are OS- and locale-dependent.