[candidate_profile] Add new candidate dashboard module to provide overview of candidates (#6092)
This adds the framework for a candidate dashboard, genericifying the work started by @zaliqarosli. It only adds the core framework so that modules can register widgets into the dashboard and new cards will be added to the appropriate modules in subsequent PRs.

Widgets registered to the dashboard consist of a React component which get wrapped in a Card component, and displayed in a CSS grid. The module passes the props of BaseURL, Candidate, Visits, and VisitMap, plus any props passed to the \LORIS\candidate_profile\CandidateWidget object returned by getWidgets to the React component. BaseURL is the LORIS base URL. Candidate is the candidate object returned by the API, Visits is an array of all visits for that candidate returned from the API, and VisitMap is a mapping of VisitLabel -> SessionID to be used for links (where appropriate) since the API doesn't expose this data. These default props are helpers to avoid redundant requests to the API, since nearly every card will need some combination of this data.

The only card provided by the module itself is a candidate info card, which summarizes the data returned by the API.