[WIKI] Install - Rewrite "Installing Loris in Depth" for new Wiki (#5253)
This rewrites the wiki page "Installing LORIS in-depth" and commits it to version control.

The wording is updated to reflect recent PRs updating the install process, such as the Ubuntu and CentOS readmes. (e.g. #5203)

The rewrite removes references to ssh and also makes clear where the install process diverges for production installs vs. development installs.

Resolves #5001