[conflict_resolver] Fix links to conflict resolver with filters set (#6132)
The conflict resolver updates the URL when filters are modified to
reflect the current state of the filters. However, when these links
are loaded on their own the filters are not reflected in the results
until some other filter is changed.

This fixes the conflict resolver so that after data is loaded the
filter automatically applies to the results.