[candidate_profile] Add Candidate Parameters to Candidate Info card (#6125)
This adds a new type of widget ('candidateinfo') which can be used by modules or projects to add information to the "Candidate Info" card on the candidate_profile page. The new type of widget can be used by a module (or project module) to add information that can be summarized as a scalar to the quick summary.

It uses the new type to add information from the candidate_parameters module that make sense as an overview for the candidate. The Participant Status, Caveat, and any 'Candidate Parameters' category variables from the parameter_candidate table are now summarized in the card.

Other information from the candidate_parameters module is complex enough to merit its own card (ie. Consent often contains multiple entries to be summarized for different types of consent, and shouldn't be in the top-level summary.)