[issue_tracker] Include all sites in dropdown for issue tracker (#6135)
The issue tracker currently limits the dropdown to non-DCC study sites.
When entering a PSCID from a site such as the DCC, this means that there
is an (unresolvable) error about the PSCID and CenterID not matching,
because the candidate's site doesn't exist in the dropdown. This
updates the issue tracker to display all sites.