[candidate_profile] Add help text for candidate profile (#6337)
This adds some help text for the candidate_profile module. The help text is somewhat vague by necessity since it isn't context sensitive and the exact set of cards shown depends on the user's permissions and LORIS install's configurations. It only explains the visit link in the Candidate Info card (which is provided by the module itself and has minimal permissions, so should always be present.)

Fixes #6206