[MRI Violations] Searches for types of problems containing newlines fail. (#6367)
The MRI violations modules builds the list of types of problems from the Reason column of the MRICandidateErrors table. When this table has reasons contains newlines, you won't be able to search for them in the MRI Violations module. Newlines should be stripped when building the types of problem list to safeguard against this.

The reason column is populated by the MRI pipeline. The pipeline should also make sure that it does not write records with reasons containing newlines (they make no sense in that context). The raisin bread data set should also be corrected to reflect this.
Testing instructions (if applicable)

    Resolves #6366