[DataFramework Filters] Check Site Permissions Filter (#6354)
Change abstraction for DataFrameworkMenu to handle "all site" permissions with different names.

This adds a new abstract function to the DataFrameworkMenu: allSitePermissionNames. This function is meant to signal to the base class that the page's provisioner can support the HasAnyPermissionOrUserSiteMatch filter. It replaces the old "useSiteFilter" abstract function. If allSitePermissionNames is null, no site filtering is used, otherwise any of the permissions will allow access to multiple sites.