[Survey] Infinite redirect (#6644)
A header location in survey is set with a variable which can be null. This has an unpredictable behaviour, and in Firefox, it seems to cause the reload of the page, where again, the header location is interpreted, causing an infinite redirection. I solved this with a check for Null before redirecting.

Also fixed some type issues (+1 on pageNum = finalpage).

    Resolves #6587