[document_repository] Permission to access module (#6646)
Changed the permission in module.class.inc to be able to access the module when you have only the permission to delete files.

When you only have permission to delete files, the tabs upload and category disappear. However, view still grants these options as per the permission description in the schema.