[issue_tracker] Improve readability of comments and history (#6138)
The comments and history were jumbled together in a way that made
it very difficult to determine what was a comment and what was an
item being changed. This updates the display in a way that is
much more readable and should make it possible to follow the conversation
of what is happening in an issue.