[Utility] Deprecate getExistingVisitLabels() (#5796)
There were two functions for getting visit labels, getExistingVisitLabels() and getVisitList().

This replaces the contents of getVisitList() with that of getExistingVisitLabels(). The former has a better name (more concise + consistent with other functions in the class), but the latter has more robust code, including the option to filter by project.