Script for data import and export (e.g. pushing to an "open science" LORIS instance) (#5334)
This pair of scripts allows one to export data from one database and import it another database, mapping PSCIDs from the former to new values. This allows for a way of transferring data into e.g. an Open Science database while giving the candidates new identifier to preserve privacy between databases.
Data Extraction

The dataExtractor.php script is used to extract data from the source database. It can extract data from either the candidate table, the session table, or from tables containing instrument data. Other tables are unsupported and the behaviour of this script when trying to extract data from other tables is undefined/almost guaranteed to have bugs.

The script writes data to project/data_export/ in CSV format. Optionally can be written elsewhere.
Data Import

The openScienceDataImport.php script is used to import data. It takes several parameters:

An operation mode, either 'candidate', 'visits', or 'instrument', corresponding to the dataExtractor tool
A CSV file mapping old/source PSCIDs to a new PSCID and new CandID for the destination database.
A source data file generated by dataExtractor.php. This file should not be renamed as the name of the MySQL table to update will be included in the name
It analyzes the two CSV files and generates SQL commands, either INSERT or UPDATE, that can be used to insert the data into the database. These commands are written to a file in project/data_import/ with a timestamp.

Candidate information will always be UPDATE statements as the candidates are expected to be created in the candidate table beforehand.
Visits information will be a mix of UPDATE or INSERT depending on whether the visit label exists
Instrument information will always be INSERTs as it is expected that instrument information is not yet populated in the DB.