[My Preferences] Change contents of email sent to user upon changing their account details (#6169)
My Preferences was sending a user's password in plaintext to their email whenever they changed it.

Edit User does this when a user has a password generated for them by the server. In this case it is acceptable because the password must be reset on login. My Preferences shares a lot of code with User Accounts so I believe it was included as a feature accidentally.

I've created a new email template notifying a user that their account details have changed.

    Resolves #6079