Add administrator contact email as a new config setting (#6101)
Adds a new configuration setting for an admin "point of contact" email. The idea is to have an email that is monitored by a developer and/or project administrator for LORIS users to contact directly from within the software.

An example of this would be when an Exception occurs in LORIS. The error page will provide a link for the user to email an admin when the error occurs. #6076

Another example will be in the server configuration check script #5965. This script will send a test email to the admin to make sure that outgoing emails are working properly on the server.