[Exceptions] Fix null user on ExceptionHandlingMiddleware
The ExceptionHandlingMiddleware uses the PageDecorationMiddleware,
and as such needs to be chained after all the prerequisites of the
PageDecorationMiddleware (ie. the lorisinstance and user) are added
to the request object.

This converts it from a MiddlewareChainer interface into a normal
middleware and explicitly adds it to the BaseRouter, instead of the
top level chain.

Replaces #6820.