Add callback for updates of filters
This adds a callback for updating filters in a FilterableDataTable.
This is intended to make it possible to add features such as context
sensitive filters, where the option selected for filter A might affect
the options available in filter B (for instance, selecting a project filter
affecting which subprojects are available in a subproject filter.)

Since changing the options might result in filter B being in an invalid
state (if the option that was selected is no longer valid when A
changes) the FilterableDataTable also removes invalid choices from being
passed down to the DataTable or Filter components upon rendering.