[timepoint_list] Hiding Candidate Info and View Imaging Datasets buttons (#6815)
This hides the "Candidate Info" and "View Imaging Datasets" buttons along with the "Create time point" button when the necessary permissions are not present. This was done with the assumption that "View Imaging Datasets" should only be visible if the user has one of the four imaging browser permissions (please correct me if this assumption is wrong). Test plan documentation was edited accordingly. If the user has none of the relevant permissions, the "Actions:" label is also hidden.

    Resolves #6673