[Database] Add return type to pselectOne() and add new pselectOneInt() (#6171)
Adds a return type of string to pselectOne() in order to assist us with using strict types throughout LORIS.

Also adds a new function pselectOneInt() as a wrapper of pselectOne(). This new function automatically validates the cell returned by pselectOne and converts it to an integer. It returns null when the query results in something that can't be reliably cast to an integer, such as the empty set.

This is to assist in cases where integer values are to be retrieved from the database, such as COUNT() results, IDs, etc. Rather than require the calling code to wrap results with intval() (which is problematic for some inputs), or int(), they can call this function to have validation be done automatically.

    Resolves #6164