[Core] Fix fatal error caused by exception handling middleware
The second argument to the "process" function is a RequestHandler,
but was incorrectly passed a response directly in the case where
the BaseRouter got a 404 error.

This is not a case that should generate any exceptions, since it's
not dependent on any user data (other than the URL from the request
object, which is just used for the error page).

Remove the ExceptionHandlingMiddleware to fix this case.