[imaging_browser] Fix inefficient query for first acquisition date
The mri_acquisition_date table dropped in LORIS 22.0, and replaced by
a JOIN to the `parameter_file` table in order to look up the header
directly. However, ths query is missing a clause to restrict it to
only look up the `acquisition_date` header, resulting in an
unconditional join to `parameter_file`. Due to the size of the table
in existing projects, this can make the imaging browser menu page
take minutes to load.

This adds a where clause to restrict it to joining the `acquisition_date`
header, making the load time of the module usable again.