[Test] Fixing unittests causing Travis failure (#6903)
The unit tests for NDB_BVL_Instrument that were merged with main in PR #6819 were causing Travis to fail. This fixes those errors.

The reason it failed Travis was because the changes made in merged PR #5352 were not added to the testing PR. These changes caused it to fail Travis but were not flagged before.