[Data Tables] Add ability for data tables to stream data (#6853)
This is a proof of concept. It loads the data for the data dictionary
module from a binary stream rather than a JSON object and displays the
data in the table as it is loading. The stream consists of rows separated
by the byte 0x1f, and columns separated by 0x1e. The entire stream is
terminated with the byte 0x04.

FieldOptions are moved to a different endpoint rather than crammed
into the JSON since there is no longer any JSON to cram it into.