[survey_account] fix can't add survey to candidate in main branch (#6932)
This fixes a number of issues created by PR #5352. The semantics
of the instrument's getSessionID was changed in that PR to
throw an exception if it could not find the session id. This
causes a number of problems when the instrument class is instantiated
without a CommentID, and as well as across modules that use the
instrument class.

This fixes the exceptions that were introduced incorrectly throughout
LORIS by changing it to return null instead of throwing an exception
when a sessionID can not be determined.

Resolves #6913
Resolves #6967
Resolves #7019