Add AccessibleResource interface to TimePoint and cleanup Candidate access logic
The candidate access logic was treating sites and projects
inconsistently in terms of whether or not you needed access
to the Registration or any TimePoint's. This implements the
AccessibleResource on a TimePoint and uses it to clean up the
logic of the Candidate's isAccessibleBy to be closer to the
intended behaviour.

Time timepoint_list now filters out any timepoints that the
user who's logged in doesn't have access to.