Remove strict equality check for in_array
The documentation for `in_array` says the third boolean
parameter makes in_array check type, but that doesn't seem
to be the case. It seems to make it check for identity (ie.
the exact same instance of the object), which means if 2
different ProjectIDs with the same value are sent the check
fails.