[Candidate] Add get last visit subproject (#5946)
This adds a new function called getLastVisitSubproject in the candidate class. This new function replaces the function getSubprojectIDUsingCandID in Utility class.

getSubprojectIDUsingCandID is only currently being called in one place, the instrument_list class. The returned value there however affects nothing. It is passed as an argument to the function getBatteryInstrumentList that then does nothing with the parameter.