[GUI] Do not separate checkboxes from linebreaks
I noticed while testing CCNA on the user_accounts page there
are often line breaks inserted between a checkbox and the
label at the end of the line when multiple checkboxes are used
in a group (because of the number of sites, the "Examiner At" spans
multiple lines while it doesn't in raisinbread/most projects.)

This ensures that there's no linebreaks between a checkbox
in LorisForm and the label that goes with it. I can't think
of any circumstances where that would be desirable behaviour.