[data_dict] Convert list of instruments back to multiselect
Remove references to "search keyword" in test plan. The
search keyword was lost in reactification in favour of filters
consistent with the data table columns, like in other modules.
(Searching across multiple columns with the same filter is not
supported by our FilterableDataTable)

Update the "Source From" column to be a multiselect. This was
also lost in reactification, but is a simple fix.

Resolves #7205