Remove PHPMD and test php8
This removes phpmd from the `make checkstatic` target and
no longer uses it in LORIS. It crashes with PHP8, preventing
us from testing/developing with the latest version of PHP over
a month after its release.

This removes it from the checkstatic pipeline and adds php8
to our continuous integration.