[NDB_Factory] Remove testing mode (#7199)
The "Testing" mode in NDB_Factory complicates our code
and frequently confuses phan. It appears to not be used
by our testing suite, and the test suite often explicitly
sets it to not-test mode in order to get real connections,
and uses setDatabase/setConfig/etc to inject mocks rather
than using the factory's internal test mode. This removes
the mode and simplifies our code.