[genomic_browser] Fix PhanUndeclaredProperty errors in genomic_browser
This fixes 2 PhanUndeclaredProperty errors in the genomic browser. The
function(s) "getPSC" are never called, and if they were to be called
would return an invalid value since they're referring to a non-existant
class property ("PSC" when the property is called "_PSC").

This fixes the errors by deleting the unused/incorrect code.