[Tests] Fix PhanUndeclaredProperty type errors in tests
This fixes the PhanUndeclaredProperty type errors in the test
directory. These are mostly caused by phan getting confused by
type juggling between mock objects and real objects. Explicitly
telling phan what type an object is at any given time mostly
fixes the issues.

Note that there is one error that remains in the NDB_Page type
for dateOptions. This will be handled separately because the
problem is with the NDB_Page class, not the tests.