[Datadict] filter issue when table name is different then test_name (#7151)
The Data Dictionary allows for filtering on the data from the parameter_type table. The parameter_type table is built by the data_dictionary_builder script which parses the LINST format of any instrument to generate the table entries. the LINST format of a PHP instrument is generated by the lorisform_parser script which writes the TABLE of the instrument into the LINST file it creates...

result: The table name is stored in the parameter_type SQL table, and the datadict module expects to filter on test name causing any filtering by instrument to return 0 results when the testname and the table name of an instrument differs.

This fixes the datadict module to work when the table name and test_name are different.