Remove NDB_BVL_Instrument_instrument_preview.class.inc
This class was at one point in time used for by the
instrument builder to render the instrument using the
same mechanisms as normal instruments, in order to give
a preview of the instrument form being built.

There is no longer any "Preview" button in the instrument_builder
module and the class only adds maintenance overhead.