[phan] Fix all calls to undeclared methods (#7309)
This fixes all calls to undeclared methods caught by phan and removes the PhanUndeclaredMethod from the suppressed issue types.

The fixes are broadly classified into:

    False positives caused by phan/phan#2628
    False positives caused by phan/phan#2129
    Mostly harmless positives caused by chaining withStatus on a PSR interface after a different call which lower the specificity from ResponseInterface to MessageInterface (because of which level of inheritance the method is defined on..)
    Real positives

The first class of problems is fixed by putting the variable into a simple variable so that @phan-var annotations can be used on them to assert the type. These should eventually be removed when the above issue is solved.
The second class of problems by asserting the correct type before the calls
The third class is solved by re-ordering the method chaining
The fourth class is fixed, mostly by adding method_exists or instanceof checks where required.